let minimal OST compile with barebone CentOS 5.5
A minimalistic OST (no info, no gfx, no gui, no img) compiles
and links with a default CentOS installation. Only eigen2 is
required as an additional dependency. But since that's header
only library that's not a problem for bundling a standalone OST.

Tested on CentOS 5.5 with boost 1.33.1 and MacOS X with boost
1.46.

+++ modules/io/tests/test_star_parser.cc
@@ -27,6 +27,8 @@
 
 #define BOOST_AUTO_TEST_DYN_LINK
 #include <boost/test/unit_test.hpp>
+#include <boost/test/auto_unit_test.hpp>
+#include <boost/test/floating_point_comparison.hpp>
 
 
 using namespace ost;
@@ -131,7 +133,7 @@ public:
                          const std::vector<StringRef>& columns) 
   {
     BOOST_CHECK_EQUAL(columns[0][0], cur_char_);
-    BOOST_CHECK_EQUAL(columns[0].size(), 1);
+    BOOST_CHECK_EQUAL(columns[0].size(), static_cast<size_t>(1));
     ++cur_char_;
     BOOST_CHECK_EQUAL(columns[1][0], cur_char_);
     BOOST_CHECK_EQUAL(columns[1].size(), 1);
